Move missingness utilities to `missing.R/c/h`
