+++ R/missing.R
@@ -1,0 +1,5 @@
+#' @export
+#' @rdname vec_equal
+vec_equal_na <- function(x) {
+  .Call(vctrs_equal_na, x)
+}
