+++ man/vec_equal.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/equal.R
+% Please edit documentation in R/equal.R, R/missing.R
 \name{vec_equal}
 \alias{vec_equal}
 \alias{vec_equal_na}
