+++ src/vctrs.h
@@ -263,10 +263,12 @@ bool vec_is_unspecified(SEXP x);
 #include "conditions.h"
 #include "dictionary.h"
 #include "dim.h"
+#include "equal.h"
 #include "hash.h"
 #include "lazy.h"
 #include "match-compare.h"
 #include "match-joint.h"
+#include "missing.h"
 #include "names.h"
 #include "order-collate.h"
 #include "order-groups.h"
